API/requirement.txt
os
io
flask
numpy
google.cloud
google.cloud.vision
mltk
nltk.stem
nltk.corpus
nltk.tokenize
gensim.parsing.preprocessing
sklearn.metrics.pairwise
transformers
torch
json